Refactored Sidebar
